Prepare handling of multiple message types (for bidding)
Partially implement the bidding feature. Actually no bidding is made yet
(end-to-end test fails), but we have extracted a class to parse the
messages received from the server.

"Disconnect on close" was implemented, to avoid conflicts when tests are
run sequentially.

Adaptations were made to use Mockito instead of jMock.

This closes Chapter 12.